backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.6",
    "pandas>=2.2.3",
    "passlib[bcrypt]>=1.7.4",
    "psycopg>=3.2.3",
    "psycopg-pool>=3.2.4",
    "pydantic-settings>=2.7.1",
    "pyjwt[crypto]>=2.10.1",
    "python-multipart>=0.0.20",
]

[dependency-groups]
dev = [
    "isort>=5.13.2",
    "ruff>=0.9.1",
]

[tool.ruff.lint]
select = ["I"]

[tool.isort]
profile = "black"
